requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.11.7
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0
python-multipart==0.0.9
python-dotenv==1.0.0
aiofiles==24.1.0
